the first assignment was to write a program to print the first 100 Fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,19 +9,19 @@
 
 int main(void)
 {
-	unsigned long int num1, num2, next;
+	double num1, num2, next;
 	int count;
 
 	num1 = 1;
 	num2 = 2;
 
-	printf("%ld, %ld, ", num1, num2);
+	printf("%.0lf, %.0lf, ", num1, num2);
 
 	for (count = 3; count <= 98; count++)
 	{
 		next = num1 + num2;
 
-		printf("%ld", next);
+		printf("%.0lf", next);
 
 		if (count < 98)
 			printf(", ");
